+++ PoloniexClient/Poloniex/PoloniexCurrencyInfo.cs
@@ -7,9 +7,9 @@ using System.Threading.Tasks;
 namespace CryptoMarketClient.Poloniex {
     public class PoloniexCurrencyInfo {
         public string Currency { get; set; }
-        public double MaxDailyWithdrawal { get; set; }
-        public double TxFee { get; set; }
-        public double MinConfirmation { get; set; }
+        public decimal MaxDailyWithdrawal { get; set; }
+        public decimal TxFee { get; set; }
+        public decimal MinConfirmation { get; set; }
         public bool Disabled { get; set; }
     }
 }
